+++ backend/src/lib.rs
@@ -9,5 +9,5 @@ pub mod startup;
 pub mod tasks;
 pub mod telemetry;
 pub mod utils;
-pub mod sse;
+pub mod ws;
 pub mod dedicated_executor;