+++ docker/run.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-
-pm2 start /app/apps.json
-/usr/sbin/nginx