+++ e2e/cypress/support/commands.ts
@@ -55,14 +55,17 @@ Cypress.Commands.add('openDrawerIfExists', () => {
 
 Cypress.Commands.add('closeDrawerIfExists', () => {
   const selector = '[data-test=closeDrawer]:visible';
-  return cy.getBySel('panesMobile').then(($body) => {
-    if ($body.find(selector).length > 0) {
-      cy.get(selector).then(($button) => {
-        $button.trigger('click');
-        cy.getBySel('drawer').should('not.exist');
-      });
-    }
-  });
+  return cy
+    .getBySel('drawer')
+    .should(Cypress._.noop)
+    .then(($drawer) => {
+      if ($drawer.find(selector).length > 0) {
+        cy.get(selector).then(($button) => {
+          $button.trigger('click', { force: true });
+          cy.getBySel('drawer').should('not.exist');
+        });
+      }
+    });
 });
 
 Cypress.Commands.add('waitForDrawerToClose', () => {
@@ -138,8 +141,47 @@ Cypress.Commands.add('changeActiveCollection', (id: string) => {
   cy.wait('@apiGetItems');
 });
 
+Cypress.Commands.add('expandCollection', (id: string) => {
+  cy.intercept({
+    method: 'PUT',
+    url: getApiPath(`preferences/expand/${id}`),
+  }).as('apiExpandCollection');
+
+  cy.openDrawerIfExists();
+  cy.getBySel(`collection-id-${id}`).within(() => {
+    cy.getBySel('chevronExpand').click();
+  });
+  cy.closeDrawerIfExists();
+
+  cy.wait('@apiExpandCollection').its('response.statusCode').should('eq', 200);
+});
+
+Cypress.Commands.add('collapseCollection', (id: string) => {
+  cy.intercept({
+    method: 'PUT',
+    url: getApiPath(`preferences/collapse/${id}`),
+  }).as('apiCollapseCollection');
+
+  cy.openDrawerIfExists();
+  cy.getBySel(`collection-id-${id}`).within(() => {
+    cy.getBySel('chevronCollapse').click();
+  });
+  cy.closeDrawerIfExists();
+
+  cy.wait('@apiCollapseCollection')
+    .its('response.statusCode')
+    .should('eq', 200);
+});
+
 Cypress.Commands.add('resetDatabaseByApi', () => {
   return cy.request('POST', `${Cypress.env('api_url')}/e2e/reset_db`);
+});
+
+Cypress.Commands.add('resetFeedPages', () => {
+  return cy.request(
+    'PUT',
+    `${Cypress.env('feeds_url_cypress')}/feed/reset_pages`
+  );
 });
 
 Cypress.Commands.add('signupByApi', (username, password, displayName) => {
@@ -184,3 +226,10 @@ Cypress.Commands.add(
     );
   }
 );
+
+Cypress.Commands.add('putFeedPageByApi', ({ feedName, page }) => {
+  return cy.request(
+    'PUT',
+    `${Cypress.env('feeds_url_cypress')}/feed/${feedName}/page/${page}`
+  );
+});